refactor: removed useless code and styles. Some styles changed
