+++ frontend/components/Primitives/Button.tsx
@@ -38,20 +38,6 @@ const Button = styled("button", {
   variants: {
     size: {
       "1": {
-        borderRadius: "$1",
-        height: "$5",
-        px: "$2",
-        fontSize: "$1",
-        lineHeight: "$sizes$5",
-      },
-      "2": {
-        borderRadius: "$2",
-        height: "$6",
-        px: "$3",
-        fontSize: "$3",
-        lineHeight: "$sizes$6",
-      },
-      "3": {
         borderRadius: "$2",
         height: "$32",
         px: "$4",
@@ -84,7 +70,6 @@ const Button = styled("button", {
       },
       blue: {
         backgroundColor: "$blue11",
-        boxShadow: "inset 0 0 0 1px $colors$blue7",
         color: "white",
         "@hover": {
           "&:hover": {
@@ -102,50 +87,6 @@ const Button = styled("button", {
           {
             backgroundColor: "$blue4",
             boxShadow: "inset 0 0 0 1px $colors$blue8",
-          },
-      },
-      green: {
-        backgroundColor: "$green2",
-        boxShadow: "inset 0 0 0 1px $colors$green7",
-        color: "$green11",
-        "@hover": {
-          "&:hover": {
-            boxShadow: "inset 0 0 0 1px $colors$green8",
-          },
-        },
-        "&:active": {
-          backgroundColor: "$green3",
-          boxShadow: "inset 0 0 0 1px $colors$green8",
-        },
-        "&:focus": {
-          boxShadow: "inset 0 0 0 1px $colors$green8, 0 0 0 1px $colors$green8",
-        },
-        '&[data-radix-popover-trigger][data-state="open"], &[data-radix-dropdown-menu-trigger][data-state="open"]':
-          {
-            backgroundColor: "$green4",
-            boxShadow: "inset 0 0 0 1px $colors$green8",
-          },
-      },
-      red: {
-        backgroundColor: "$loContrast",
-        boxShadow: "inset 0 0 0 1px $colors$slate7",
-        color: "$red11",
-        "@hover": {
-          "&:hover": {
-            boxShadow: "inset 0 0 0 1px $colors$slate8",
-          },
-        },
-        "&:active": {
-          backgroundColor: "$red3",
-          boxShadow: "inset 0 0 0 1px $colors$red8",
-        },
-        "&:focus": {
-          boxShadow: "inset 0 0 0 1px $colors$red8, 0 0 0 1px $colors$red8",
-        },
-        '&[data-radix-popover-trigger][data-state="open"], &[data-radix-dropdown-menu-trigger][data-state="open"]':
-          {
-            backgroundColor: "$red4",
-            boxShadow: "inset 0 0 0 1px $colors$red8",
           },
       },
       transparentWhite: {
@@ -223,105 +164,6 @@ const Button = styled("button", {
       },
     },
   },
-  compoundVariants: [
-    {
-      variant: "gray",
-      ghost: "true",
-      css: {
-        backgroundColor: "transparent",
-        color: "$hiContrast",
-        "@hover": {
-          "&:hover": {
-            backgroundColor: "$slateA3",
-            boxShadow: "none",
-          },
-        },
-        "&:active": {
-          backgroundColor: "$slateA4",
-        },
-        "&:focus": {
-          boxShadow: "inset 0 0 0 1px $colors$slateA8, 0 0 0 1px $colors$slateA8",
-        },
-        '&[data-radix-popover-trigger][data-state="open"], &[data-radix-dropdown-menu-trigger][data-state="open"]':
-          {
-            backgroundColor: "$slateA4",
-            boxShadow: "none",
-          },
-      },
-    },
-    {
-      variant: "blue",
-      ghost: "true",
-      css: {
-        backgroundColor: "transparent",
-        "@hover": {
-          "&:hover": {
-            backgroundColor: "$blueA3",
-            boxShadow: "none",
-          },
-        },
-        "&:active": {
-          backgroundColor: "$blueA4",
-        },
-        "&:focus": {
-          boxShadow: "inset 0 0 0 1px $colors$blueA8, 0 0 0 1px $colors$blueA8",
-        },
-        '&[data-radix-popover-trigger][data-state="open"], &[data-radix-dropdown-menu-trigger][data-state="open"]':
-          {
-            backgroundColor: "$blueA4",
-            boxShadow: "none",
-          },
-      },
-    },
-    {
-      variant: "green",
-      ghost: "true",
-      css: {
-        backgroundColor: "transparent",
-        "@hover": {
-          "&:hover": {
-            backgroundColor: "$greenA3",
-            boxShadow: "none",
-          },
-        },
-        "&:active": {
-          backgroundColor: "$greenA4",
-        },
-        "&:focus": {
-          boxShadow: "inset 0 0 0 1px $colors$greenA8, 0 0 0 1px $colors$greenA8",
-        },
-        '&[data-radix-popover-trigger][data-state="open"], &[data-radix-dropdown-menu-trigger][data-state="open"]':
-          {
-            backgroundColor: "$greenA4",
-            boxShadow: "none",
-          },
-      },
-    },
-    {
-      variant: "red",
-      ghost: "true",
-      css: {
-        backgroundColor: "transparent",
-        "@hover": {
-          "&:hover": {
-            backgroundColor: "$redA3",
-            boxShadow: "none",
-          },
-        },
-        "&:active": {
-          backgroundColor: "$redA4",
-        },
-        "&:focus": {
-          boxShadow: "inset 0 0 0 1px $colors$redA8, 0 0 0 1px $colors$redA8",
-        },
-        '&[data-radix-popover-trigger][data-state="open"], &[data-radix-dropdown-menu-trigger][data-state="open"]':
-          {
-            backgroundColor: "$redA4",
-            boxShadow: "none",
-          },
-      },
-    },
-  ],
   defaultVariants: {
     size: "1",
     variant: "gray",
