+++ frontend/pages/_app.tsx
@@ -3,7 +3,7 @@ import Head from "next/head";
 import { IdProvider } from "@radix-ui/react-id";
 import globalStyles from "../styles/globals";
 import Layout from "../components/Layout/Layout";
-import { ContextProvider } from "../store/store-context";
+import { ContextProvider } from "../store/store";
 
 function App({ Component, pageProps }: AppProps): JSX.Element {
   globalStyles();
